sean-otoole
The latest working architecture includes an encoder and three heads, each passing a loss functions back to the encoder, corresponding to patient outcome, reconstruction, and technology (microarray vs sequencing) loss. A gradient reversal layer is used for the technology backpropagation. The encoder is trained to contain a faithful representation of the input, emphasize information relating to patient outcome and reduce the presence of domain specific information.